requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.22
django-filter==23.4
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.2
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
